Ik ga pushen

+++ Tower Defence/Assets/Scripts/Currency/Currency.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 
 public class Currency : MonoBehaviour {
-	public int myCurrency;
+	public float myCurrency;
 }
